requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
react-vite
axios
requests
requests-oauthlib
python-openid
django-allauth

